# swaggerDoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "Teamwork is an online platform where people from same organization meet for social benefits.",
    "version": "1",
    "title": "Auto-Mart",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "conquerorsword36@yahoo.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/api/v1/article/:articleId": {
      "patch": {
        "tags": [
          "Article"
        ],
        "summary": "Update/edit an article",
        "description": "",
        "operationId": "updateArticle",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ArticleId",
            "type": "string",
            "description": "Mark Article Ad as sold",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "operation successful",
            "schema": {
              "$ref": "#/definitions/Article"
            }
          },
          "400": {
            "description": "Invalid Input"
          },
          "404": {
            "description": "Article not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Article"
        ],
        "summary": "delete an article",
        "description": "Multiple status values can be provided with comma separated strings",
        "operationId": "Article_Id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "article_Id",
            "in": "req.body",
            "description": "id of article to be deleted",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "Required",
            "description": "required field",
            "schema": {
              "properties": {
                "price": {
                  "type": "success",
                  "description": "Article successfully deleted"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Article"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "/api/v1/feed": {
        "get": {
          "tags": [
            "article"
          ],
          "summary": "View an Articles in descending other",
          "description": "",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/CarArray"
              }
            },
            "404": {
              "description": "Invalid input"
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
        }
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "token",
      "in": "header"
    }
  },
  "definitions": {
...
    "Article": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "article_id": {
          "type": "integer",
          "format": "uuid"
        },
        "created_on": {
          "type": "string",
          "example": "user id"
        },
        "title": {
          "type": "string",
          "example": "New Cloths"
        },
        "article": {
          "type": "string",
          "example": "Benz"
        },
        "author_id": {
          "type": "string",
          "format":"uuid"
        }
      }
    },
...
  },
...
}